# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des catégories",
        "description": "Documentation de l'API de gestion des catégories",
        "contact": {
            "email": "sarah.mourlevat@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Display a listing of the resource.",
                "operationId": "54ca80195310a25e51af8912a17cf3d1",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Products"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Store a newly created resource in storage.",
                "operationId": "84771ac90926cf4f51cf3328d98084a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "stock",
                                    "categories"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Productname"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "12.25"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": "1"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Products"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "description": "Display the specified resource.",
                "operationId": "1196773a65ac8616e64c34ae4f5e8218",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Products"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a specific product",
                "description": "Update the specified resource in storage.",
                "operationId": "e67dc9715fd5d4aca7ea3ecbf574982a",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Productname"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "12.25"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": "1"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Products"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a specific product",
                "description": "Remove the specified resource from storage.",
                "operationId": "19de1f9e3f99b025e56208bf1e2f4047",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Categories": {
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "title": {
                        "description": "The title of the category",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the category",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The timestamp when the user was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "The timestamp when the user was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "Products": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "name": {
                        "description": "The name of the product",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the product",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "The stock of the product",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "The timestamp when the user was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "The timestamp when the user was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T12:34:56Z"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Categories"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}